extracted RealTable to its own file
